refactor: rename to TagContextMenu
